move schema link actions helper
