+++ lib/shared/actions/page-builder.js
@@ -2,7 +2,7 @@ import actionTypes from 'actions';
 import forEach from 'lodash/forEach';
 import getContextFragment from 'helpers/page-builder/get-context-fragment';
 import getElement from 'helpers/page-builder/get-element';
-import getSchemaLinkActions from 'helpers/schema-link-actions';
+import getSchemaLinkActions from 'helpers/page-builder/schema-link-actions';
 import merge from 'lodash/merge';
 import pageBuilderElements from 'elements';
 import {dropDraft, saveDraft, updateDraft} from 'actions/draft';
