+++ lib/shared/helpers/page-builder/schema-link-actions.js
